Implemented test code
